--- packages/react-next/src/components/Dropdown/index.ts
@@ -1,3 +1,0 @@
-export * from './Dropdown';
-export * from './Dropdown.base';
-export * from './Dropdown.types';
